Revert "Merge branch 'je/RCORE-2094' into nc/skunkworks_perf_query_compression"
This reverts commit 4dd54838f03e42e23d4dbcc00018c7bd4156086b, reversing
changes made to 3acd2789725351ba54ab1a6b1ffcc24f1b1d8671.

+++ src/realm/integer_flex_compressor.cpp
@@ -38,7 +38,7 @@ void FlexCompressor::init_header(char* h, uint8_t flags, uint8_t v_width, uint8_
 }
 
 void FlexCompressor::copy_data(const Array& arr, const std::vector<int64_t>& values,
-                               const std::vector<unsigned>& indices)
+                               const std::vector<size_t>& indices)
 {
     using Encoding = NodeHeader::Encoding;
     REALM_ASSERT_DEBUG(arr.is_attached());
